+++ packages/core/index.ts
@@ -1,5 +1,5 @@
 import { defineAdderConfig, defineAdderTests, defineAdder, defineAdderOptions, defineAdderChecks } from "./adder/config.js";
-import { executeCli } from "./utils/common.js";
+import { executeCli } from "./utils/cli.js";
 import dedent from "dedent";
 
 export { defineAdderConfig, defineAdder, defineAdderTests, defineAdderOptions, defineAdderChecks, executeCli, dedent };
